// src/app/models/master/masWebsiteBannerData.ts
export interface MasWebsiteBannerData {
     bannerId?: number;
     first?: number;

	 bannerName?: string;
	 bannerLink?: string;
	 displayStartDate?: string;
	 displayEndDate?: string;
	 bannerImageTh?: string;
	 bannerImageEn?: string;
	 orderBy?: number;
	 activeFlag?: boolean;
     lastUpDateList?: Date;

     mode?: string;
     rowNum?: number;
     size?: number;

	 filename?: string;
	 prefix?: string;
	 module?: number;
     base64?: string;

     fullNameTh?: string;
     fullNameEn?: string;
}